Move xmlpull api module into top-level dir
