+++ xmlpull/src/test/java/org/xmlpull/v1/tests/TestSerialize.java
