+++ xmlpull/src/test/java/org/xmlpull/v1/tests/TestSimple.java
