+++ xmlpull/src/test/resources/simple.xml
