+++ xmlpull/src/test/resources/simple2.xml
