+++ xmlpull/src/test/resources/typical.xml
