--- xmlpull/xmlpull/.gitignore
@@ -1,11 +1,0 @@
-# Maven's build directory and a handy local folder
-target/
-local/
-
-# Eclipse, Netbeans and IntelliJ files
-/.*
-!.gitignore
-/nbproject
-/*.ipr
-/*.iws
-/*.iml
